refactor: export request functions to lib
